Update facebbok API version to 11.0
The previous version, 10.0 got deprecated, so I'm manully bumping this to 11.0 as a workaround.
This isn't meant to be a final solution we will need to eventually reconcile back with upstream, or implement our our integrations layer.
But as this is currently blocking customers and sales we wanna make sure to get around it sooner than later