Fixup ruby-3.5 compatibility
`irb`, `logger` & `ostruct` are used by SDK and was moved from default
to bundled gems in ruby-3.5 [0]. Add them to the gemspec to fix
compatibility.

We're using `head` in CI instead of `3.5` because it's closer to the
final version, see [1]

No `continue-on-error` because we're only testing `gem build` for now,
which should work with `head` no matter what. Could be skipped later
if needed.

Fixes #220

[0] https://bugs.ruby-lang.org/issues/20309
[1] https://github.com/ruby/ruby-builder/pull/4#issuecomment-717885932